Improved Agent and Model classes and interaction, DDPG, DQN
Fixed the construction of Agent class and sublcass on all levels.
Implemented the public functions in Agent with common behavior and made them
call private functions that are not implemented in the base class.
Same for Model class. Made the appropriate DDPG changes. Added DQN model
as well as agent.