Add env wrapper for stacking frames and adapt replay buffer
New env wrapper that stacks frames and returns np.array. For image observations,
on the outside, the Replay Buffer now behaves as if it stores the whole state,
but it only stores the last observation. Tuples of transitions can now be stored
in one function call. Memory efficiency is preserved.