Fix episode id counters and rewards when using max episode steps
Major fix in StatsRecorder - episode_id is now incremented every time
env.reset() is called instead of tracking done. Previously, this caused problems
in Atari environments based on lives. Specifically, for the agent end of life ==
end of episode. However, for StatsRecorder the episode continued since the rewards
are supposed to be recorded over the whole bunch of lives as provided by the ALE.
Fix a bug in max episode steps - the corresponding wrapped must be applied
before the Monitor, otherwise the episode reward was wrong.
Minor fix in setting the mode of the Monitor