Update evaluation procedure and reports, clearly separate train and eval
Big change in StatsRecorder - it now tracks both 100 episodes mean and scores
from single evaluation runs. All data is properly saved so plots can be reconstructed.
Only single mode of operation is allowed for the whole life of StatsRecorder.
Restoring statistics is improved. Save and restore does not mess up any of the
stats. Proper train steps are used on save and restore. Eval stats are reported only
at the end of the run.
Small bug fix in Agent which did not continue eval run if process was killed during
eval. Now Agent saves the model variables from the best eval run. (A small bug when
there has been a restore before - fix is on the way).
Eval episodes in DQN are limited to 108K frames. Different wrappers for train and
eval envs.